requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
pandas==2.0.3
PyYAML==6.0.1